Refactored code. Organized files into folders. Created single TaskBase clas
