+++ lithuanian-language-learning-tool/Components/Pages/SpellingTaskPage.razor
@@ -1,0 +1,143 @@
+﻿@page "/spellingtask"
+@using Models
+@inherits SpellingTaskBase
+
+<h1>Lietuvių rašybos užduotys</h1>
+
+@if (!startExercise)
+{
+    <TaskUploadPrompt OnTaskUploaded="LoadCustomTasks" OnUseDefaultTasks="StartWithDefaultTasks" />
+}
+else
+{
+    @if (!showSummary)
+    {
+        
+        
+        @if (reviewMode)
+        {
+            <TaskReview Task="currentTask" OnBackToSummary="() => showSummary = true" />
+        }
+        else 
+        {
+            <Timer SecondsToRun="10" TimerOut="TimerOut" />
+
+            <div class="task-box" style="height: 300px;">
+                <textarea disabled @bind=currentTask.Sentence class="edit-textarea"></textarea>
+            </div>
+
+        
+            <OptionsSection Task="currentTask" TTask="SpellingTask" OnOptionSelected="CheckAnswer" />
+        
+
+            if (showFeedback)
+            {
+                <div class="feedback @(currentTask.TaskStatus ? "correct" : "incorrect")">
+                    @feedbackMessage
+                </div>
+            }
+
+            <button class="next-button" @onclick="NextTask">Kita užduotis</button>
+        }
+
+        
+    }
+    else
+    {
+        <SummaryPage 
+            CorrectAnswersCount="correctAnswersCount"
+            TTask="SpellingTask"
+            Tasks="tasks" 
+            Score="score" 
+            OnTaskSelected="GoToTask" 
+            OnRestart="StartExercise" />
+    }
+}
+
+
+
+@* @page "/spellingtask"
+@inherits SpellingTaskBase
+
+@if(!startExercise)
+{
+    <TaskUploadPrompt OnTaskUploaded="LoadCustomTasks" OnUseDefaultTasks="StartWithDefaultTasks" />
+}
+else
+{
+
+    @if (!showSummary)
+    {
+        <h1>Lietuvių Rašybos Užduotys</h1>
+        @if (!reviewMode)
+        {
+            <div class="score-label">
+                <h2>Taškai: @score</h2>
+            </div>
+
+            <Timer @ref="timer" SecondsToRun="10" TimerOut="TimerOut" />
+        }
+
+
+        <div class="task-box" style="height: 300px;">
+            <textarea disabled @bind=tasks[currentTaskIndex].Sentence class="edit-textarea"></textarea>
+        </div>
+
+
+
+        @if (!reviewMode)
+        {
+            <div class="options">
+                @foreach (var option in tasks[currentTaskIndex].Options)
+                {
+                    <button class="option" @onclick="showFeedback ? null : () => CheckAnswer(option)">@option</button>
+                }
+            </div>
+        }
+
+        <!-- Explanation for revisited task -->
+        @if (reviewMode)
+        {
+            <div class="answer-section">
+                <h4>Teisingas atsakymas</h4>
+                <p class="explanation-message">@correctAnswer</p>
+            </div>
+            <div class="answer-section">
+                <h4>Paaiškinimas</h4>
+                <p class="explanation-message">@explanationMessage</p>
+            </div>
+            <button class="next-button" @onclick="() => showSummary = true">Grįžti į apžvalgą</button>
+        }
+        else if (showFeedback)
+        {
+            <div class="feedback @(isCorrect ? "correct" : "incorrect")">
+                @feedbackMessage
+            </div>
+
+            <button class="next-button" @onclick="NextTask">Kita užduotis</button>
+        }
+    }
+    else
+    {
+        <h1 class="overview-title">Užduoties apžvalga</h1>
+        <div class="container">
+            <div class="container-bright-long">
+                <div class="d-flex justify-content-between align-items-center">
+                    <p class="correct-answers">Teisingi atsakymai<br />
+                        @correctTotal / @tasks.Count</p>
+                    <p class="correct-answers">Surinkta Taškų<br /> @score</p>
+                </div>
+            </div>
+
+                @for (int i = 0; i < tasks.Count; i++)
+                {
+                    var index = i;
+                    <button class="summary-button @(taskStatus[i] ? "green" : "red")" @onclick="() => GoToTask(index)">
+                        @(i + 1)
+                    </button>
+                }
+
+        </div>
+        <button class="next-button" @onclick="RestartTasks">Bandyti dar kartą</button>
+    }
+} *@
