+++ lithuanian-language-learning-tool/Components/TaskReview.razor
@@ -1,0 +1,19 @@
+﻿@using Models
+<!-- TaskReview.razor -->
+<div class="task-box" style="height: 300px;">
+    <textarea readonly class="edit-textarea">@Task.UserText</textarea>
+</div>
+<div class="answer-section">
+    <h4>Teisingas atsakymas</h4>
+    <p class="explanation-message">@Task.CorrectAnswer</p>
+</div>
+<div class="answer-section">
+    <h4>Paaiškinimas</h4>
+    <p class="explanation-message">@Task.Explanation</p>
+</div>
+<button class="next-button" @onclick="OnBackToSummary">Grįžti į apžvalgą</button>
+
+@code {
+    [Parameter] public CustomTask Task { get; set; }
+    [Parameter] public EventCallback OnBackToSummary { get; set; }
+}
