+++ lithuanian-language-learning-tool/Helpers/CustomTaskExtensions.cs
@@ -1,3 +1,4 @@
+using lithuanian_language_learning_tool.Models;
 namespace lithuanian_language_learning_tool.Helpers
 {
     public static class CustomTaskExtensions
