+++ lithuanian-language-learning-tool/Models/SpellingTask.cs
@@ -1,0 +1,7 @@
+﻿namespace lithuanian_language_learning_tool.Models
+{
+    public class SpellingTask : CustomTask
+    {
+        // Maybe add specific Spelling Task implementation someday...
+    }
+}
