--- lithuanian-language-learning-tool/wwwroot/PunctuationTest.json
@@ -1,86 +1,0 @@
-[
-  {
-    "Sentence": "Lietuva yra nepriklausoma šalis.",
-    "Options": [".", ",", ":", ";", "!", "?"],
-    "CorrectAnswer": "Lietuva yra nepriklausoma šalis.",
-    "Explanation": "Šis sakinys yra teisingai parašytas, nereikia jokių papildomų skyrybos ženklų."
-  },
-  {
-    "Sentence": "Sniegas krenta nuo stogo ar kiemo.",
-    "Options": [",", ";", ":", ".", "!"],
-    "CorrectAnswer": "Sniegas krenta nuo stogo, ar kiemo.",
-    "Explanation": "Kablelis naudojamas atskirti dvi sakinio dalis, kurios yra sujungtos jungtuku 'ar'."
-  },
-  {
-    "Sentence": "Šiandien oras labai šiltas ir saulėtas",
-    "Options": [".", ":", ";", ",", "!"],
-    "CorrectAnswer": "Šiandien oras labai šiltas ir saulėtas.",
-    "Explanation": "Sakinys užbaigiamas tašku, nes tai yra paprastas teiginys."
-  },
-  {
-    "Sentence": "Kai kurie žmonės sako kad rytoj lis",
-    "Options": [",", ".", ";", ":", "!"],
-    "CorrectAnswer": "Kai kurie žmonės sako, kad rytoj lis.",
-    "Explanation": "Kablelis naudojamas po frazės 'sako', nes po jos eina šalutinė frazė."
-  },
-  {
-    "Sentence": "Petriukas sako kad jis negalėjo padaryti namų darbų",
-    "Options": [",", ".", ";", ":", "!"],
-    "CorrectAnswer": "Petriukas sako, kad jis negalėjo padaryti namų darbų.",
-    "Explanation": "Kablelis reikalingas po 'sako', nes po to eina šalutinis sakinys."
-  },
-  {
-    "Sentence": "Mes žinome jog kalnas yra aukštas",
-    "Options": [",", ".", ":", ";", "!"],
-    "CorrectAnswer": "Mes žinome, jog kalnas yra aukštas.",
-    "Explanation": "Kablelis reikalingas po 'žinome', nes po to eina šalutinis sakinys."
-  },
-  {
-    "Sentence": "Kada ateis vasara žmonės pradės keliauti daugiau",
-    "Options": [",", ".", ":", ";", "?"],
-    "CorrectAnswer": "Kada ateis vasara, žmonės pradės keliauti daugiau.",
-    "Explanation": "Kablelis naudojamas po frazės 'Kada ateis vasara', nes tai yra laiko šalutinė dalis."
-  },
-  {
-    "Sentence": "Lietuvoje žiemos dažnai būna labai šaltos ir vėjuotos",
-    "Options": [".", ",", ":", ";", "!"],
-    "CorrectAnswer": "Lietuvoje žiemos dažnai būna labai šaltos ir vėjuotos.",
-    "Explanation": "Sakinys užbaigiamas tašku, nes tai yra paprastas teiginys."
-  },
-  {
-    "Sentence": "Šiandien oras labai šiltas ir saulėtas",
-    "Options": [".", ":", ";", ",", "!"],
-    "CorrectAnswer": "Šiandien oras labai šiltas ir saulėtas.",
-    "Explanation": "Sakinys užbaigiamas tašku, nes tai yra paprastas teiginys."
-  },
-  {
-    "Sentence": "Kai kurie žmonės sako kad rytoj lis",
-    "Options": [",", ".", ";", ":", "!"],
-    "CorrectAnswer": "Kai kurie žmonės sako, kad rytoj lis.",
-    "Explanation": "Kablelis naudojamas po frazės 'sako', nes po jos eina šalutinė frazė."
-  },
-  {
-    "Sentence": "Petriukas sako kad jis negalėjo padaryti namų darbų",
-    "Options": [",", ".", ";", ":", "!"],
-    "CorrectAnswer": "Petriukas sako, kad jis negalėjo padaryti namų darbų.",
-    "Explanation": "Kablelis reikalingas po 'sako', nes po to eina šalutinis sakinys."
-  },
-  {
-    "Sentence": "Mes žinome jog kalnas yra aukštas",
-    "Options": [",", ".", ":", ";", "!"],
-    "CorrectAnswer": "Mes žinome, jog kalnas yra aukštas.",
-    "Explanation": "Kablelis reikalingas po 'žinome', nes po to eina šalutinis sakinys."
-  },
-  {
-    "Sentence": "Kada ateis vasara žmonės pradės keliauti daugiau",
-    "Options": [",", ".", ":", ";", "?"],
-    "CorrectAnswer": "Kada ateis vasara, žmonės pradės keliauti daugiau.",
-    "Explanation": "Kablelis naudojamas po frazės 'Kada ateis vasara', nes tai yra laiko šalutinė dalis."
-  },
-  {
-    "Sentence": "Lietuvoje žiemos dažnai būna labai šaltos ir vėjuotos",
-    "Options": [".", ",", ":", ";", "!"],
-    "CorrectAnswer": "Lietuvoje žiemos dažnai būna labai šaltos ir vėjuotos.",
-    "Explanation": "Sakinys užbaigiamas tašku, nes tai yra paprastas teiginys."
-  }
-]
